API/pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = "Network Visualization API with chat functionality and NetworkX integration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "networkx>=3.4.2",
    "openai==1.3.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "numpy>=2.2.5",
    "scipy>=1.12.0",
    "httpx>=0.27.0",
    "anyio==3.6.2",
    "starlette>=0.31.1",
    "pydantic>=2.0.0",
]
